+++ vck-openid-ktor/src/commonTest/kotlin/at/asitplus/wallet/lib/ktor/openid/TestUtils.kt
@@ -8,11 +8,12 @@ import at.asitplus.openid.OAuth2AuthorizationServerMetadata
 import at.asitplus.openid.OidcUserInfo
 import at.asitplus.openid.OidcUserInfoExtended
 import at.asitplus.openid.PushedAuthenticationResponseParameters
-import at.asitplus.openid.TokenIntrospectionJwtResponse
 import at.asitplus.openid.TokenIntrospectionResponse
-import at.asitplus.openid.TokenIntrospectionResult
+import at.asitplus.openid.TokenIntrospectionResponseJson
+import at.asitplus.openid.TokenIntrospectionResponseJwt
 import at.asitplus.openid.TokenResponseParameters
 import at.asitplus.signum.indispensable.CryptoPublicKey
+import at.asitplus.signum.indispensable.josef.JwsCompactTyped
 import at.asitplus.signum.indispensable.josef.io.joseCompliantSerializer
 import at.asitplus.wallet.eupid.EU_PID_DOCTYPE
 import at.asitplus.wallet.eupidsdjwt.EU_PID_SD_JWT_VCT
@@ -24,6 +25,7 @@ import at.asitplus.wallet.lib.data.AttributeIndex
 import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation.*
 import at.asitplus.wallet.lib.data.CredentialRepresentation
 import at.asitplus.wallet.lib.data.CredentialScheme
+import at.asitplus.wallet.lib.data.IntrospectionJwt
 import at.asitplus.wallet.lib.data.IsoMdocCredentialScheme
 import at.asitplus.wallet.lib.data.MediaTypes
 import at.asitplus.wallet.lib.data.SdJwtCredentialScheme
@@ -189,18 +191,17 @@ object TestUtils {
         headers = headersOf(HttpHeaders.ContentType, ContentType.Application.Json.toString())
     )
 
-    fun MockRequestHandleScope.respond(result: TokenIntrospectionResult): HttpResponseData = when (result) {
-        is TokenIntrospectionResponse -> respond(result)
-        is TokenIntrospectionJwtResponse -> respond(
-            joseCompliantSerializer.encodeToString<TokenIntrospectionJwtResponse>(result),
+    fun MockRequestHandleScope.respond(result: TokenIntrospectionResponse): HttpResponseData = when (result) {
+        is TokenIntrospectionResponseJson -> respond(
+            joseCompliantSerializer.encodeToString(result),
             headers = headersOf(HttpHeaders.ContentType, ContentType.Application.Json.toString())
         )
-    }
 
-    fun MockRequestHandleScope.respond(result: TokenIntrospectionResponse): HttpResponseData = respond(
-        joseCompliantSerializer.encodeToString(result),
-        headers = headersOf(HttpHeaders.ContentType, ContentType.Application.Json.toString())
-    )
+        is TokenIntrospectionResponseJwt -> respond(
+            result.value.toString(),
+            headers = headersOf(HttpHeaders.ContentType, ContentType.Application.IntrospectionJwt.toString())
+        )
+    }
 
     fun MockRequestHandleScope.respond(result: JsonObject): HttpResponseData = respond(
         joseCompliantSerializer.encodeToString(result),